requirements.txt
boto==2.49.0
boto3==1.9.130
botocore==1.12.130
bz2file==0.98
certifi==2019.3.9
chardet==3.0.4
Cython==0.29.6
docutils==0.14
idna==2.8
jmespath==0.9.4
nltk==3.4
numpy==1.16.2
pandas==0.24.2
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.0
urllib3==1.24.1
